Fix: user 도메인의 data class들의 기능별 폴더 정리 및 gson 컨버터에 대응하는 코드 수정
