--- app/src/main/java/com/example/farmus_application/model/FarmList.kt
@@ -1,5 +1,0 @@
-package com.example.farmus_application.model
-
-data class FarmList(
-    val farm_list: List<FarmDetail>
-)
