--- app/src/main/java/com/example/farmus_application/model/FarmSearch.kt
@@ -1,6 +1,0 @@
-package com.example.farmus_application.model
-
-
-import com.google.gson.annotations.SerializedName
-
-class FarmSearch : ArrayList<FarmSearchItem>()
