--- app/src/main/java/com/example/farmus_application/model/user/FindAccountReq.kt
@@ -1,9 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FindAccountReq(
-    val name: String,
-    val phoneNumber : String
-)
