--- app/src/main/java/com/example/farmus_application/model/user/FindAccountRes.kt
@@ -1,8 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FindAccountRes(
-    val email: String
-)
