--- app/src/main/java/com/example/farmus_application/model/user/FindPasswordReq.kt
@@ -1,8 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FindPasswordReq(
-    val email : String
-)
