--- app/src/main/java/com/example/farmus_application/model/user/FindPasswordRes.kt
@@ -1,8 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FindPasswordRes(
-    val password : String
-)
