--- app/src/main/java/com/example/farmus_application/model/user/LoginReq.kt
@@ -1,9 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class LoginReq(
-    val email: String,
-    val password: String
-)
