--- app/src/main/java/com/example/farmus_application/model/user/LoginRes.kt
@@ -1,11 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class LoginRes(
-    val code: Int,
-    val isSuccess: Boolean,
-    val message: String,
-    val result: LoginResult
-)
