--- app/src/main/java/com/example/farmus_application/model/user/LoginResult.kt
@@ -1,13 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class LoginResult(
-    val accesstoken: String,
-    val email: String,
-    val name: String,
-    val nickName: String,
-    val role: String,
-    val status: Boolean
-)
