--- app/src/main/java/com/example/farmus_application/model/user/ResultRes.kt
@@ -1,5 +1,0 @@
-package com.example.farmus_application.model.user
-
-data class ResultRes(
-    val result: Boolean
-)
