--- app/src/main/java/com/example/farmus_application/model/user/SignUpResult.kt
@@ -1,9 +1,0 @@
-package com.example.farmus_application.model.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SignUpResult(
-    val email: String,
-    val role: String
-)
