--- app/src/main/java/com/example/farmus_application/model/user/SignUpVerificationReq.kt
@@ -1,5 +1,0 @@
-package com.example.farmus_application.model.user
-
-data class SignUpVerificationReq(
-    val phoneNumber: String
-)
