--- app/src/main/java/com/example/farmus_application/model/user/SignUpVerificationRes.kt
@@ -1,7 +1,0 @@
-package com.example.farmus_application.model.user
-
-data class SignUpVerificationRes(
-    val code: Int,
-    val message: String,
-    val result: Boolean
-)
