--- app/src/main/java/com/example/farmus_application/model/user/VerificationReq.kt
@@ -1,7 +1,0 @@
-package com.example.farmus_application.model.user
-
-data class VerificationReq(
-    val name: String?,
-    val phoneNumber: String,
-    val usercode: String
-)
