--- app/src/main/java/com/example/farmus_application/model/user/VerificationRes.kt
@@ -1,11 +1,0 @@
-package com.example.farmus_application.model.user
-
-import java.io.Serializable
-
-data class VerificationRes(
-    val code: Int,
-    val email: String,
-    val message: String,
-    val name: String,
-    val result: Boolean
-):Serializable
