+++ app/src/main/java/com/example/farmus_application/model/user/findaccount/FindAccountRes.kt
@@ -1,0 +1,7 @@
+package com.example.farmus_application.model.user.findaccount
+
+import com.google.gson.annotations.SerializedName
+
+data class FindAccountRes(
+    @SerializedName("result") val result: Boolean
+)
