+++ app/src/main/java/com/example/farmus_application/model/user/findpassword/FindPasswordRes.kt
@@ -1,0 +1,7 @@
+package com.example.farmus_application.model.user.findpassword
+
+import com.google.gson.annotations.SerializedName
+
+data class FindPasswordRes(
+    @SerializedName("result") val result: Boolean
+)
