+++ app/src/main/java/com/example/farmus_application/model/user/login/LoginReq.kt
@@ -1,0 +1,8 @@
+package com.example.farmus_application.model.user.login
+
+import com.google.gson.annotations.SerializedName
+
+data class LoginReq(
+    @SerializedName("email") val email: String,
+    @SerializedName("password") val password: String
+)
