+++ app/src/main/java/com/example/farmus_application/model/user/signup/SignUpRes.kt
@@ -1,0 +1,10 @@
+package com.example.farmus_application.model.user.signup
+
+import com.google.gson.annotations.SerializedName
+
+data class SignUpRes(
+    @SerializedName("code") val code: Int,
+    @SerializedName("isSuccess") val isSuccess: Boolean,
+    @SerializedName("message") val message: String,
+    @SerializedName("result") val result: SignUpResult
+)
