+++ app/src/main/java/com/example/farmus_application/model/user/signup/SignUpResult.kt
@@ -1,0 +1,8 @@
+package com.example.farmus_application.model.user.signup
+
+import com.google.gson.annotations.SerializedName
+
+data class SignUpResult(
+    @SerializedName("email") val email: String,
+    @SerializedName("role") val role: String
+)
