+++ app/src/main/java/com/example/farmus_application/model/user/signup_verification/SignUpVerificationReq.kt
@@ -1,0 +1,7 @@
+package com.example.farmus_application.model.user.signup_verification
+
+import com.google.gson.annotations.SerializedName
+
+data class SignUpVerificationReq(
+    @SerializedName("phoneNumber") val phoneNumber: String
+)
