+++ app/src/main/java/com/example/farmus_application/model/user/signup_verification/SignUpVerificationRes.kt
@@ -1,0 +1,9 @@
+package com.example.farmus_application.model.user.signup_verification
+
+import com.google.gson.annotations.SerializedName
+
+data class SignUpVerificationRes(
+    @SerializedName("code") val code: Int,
+    @SerializedName("message") val message: String,
+    @SerializedName("result") val result: Boolean
+)
