+++ app/src/main/java/com/example/farmus_application/model/user/verification/VerificationRes.kt
@@ -1,0 +1,11 @@
+package com.example.farmus_application.model.user.verification
+
+import com.google.gson.annotations.SerializedName
+
+data class VerificationRes(
+    @SerializedName("code") val code: Int,
+    @SerializedName("email") val email: String,
+    @SerializedName("message") val message: String,
+    @SerializedName("name") val name: String,
+    @SerializedName("result") val result: Boolean
+)
