+++ app/src/main/java/com/example/farmus_application/model/user/withdrawal/WithdrawalRes.kt
@@ -1,0 +1,7 @@
+package com.example.farmus_application.model.user.withdrawal
+
+import com.google.gson.annotations.SerializedName
+
+data class WithdrawalRes(
+    @SerializedName("result") val result: Boolean
+)
