--- app/src/main/java/com/example/farmus_application/repository/farm/FarmDetailDataSource.kt
@@ -1,9 +1,0 @@
-package com.example.farmus_application.repository.farm
-
-import com.example.farmus_application.model.FarmDetail
-
-interface FarmDetailDataSource {
-
-    // suspend 함수 작성
-//    suspend fun getFarmList(): List<FarmDetail>
-}
