--- app/src/main/java/com/example/farmus_application/repository/farm/FarmDetailRemoteDataSource.kt
@@ -1,12 +1,0 @@
-package com.example.farmus_application.repository.farm
-
-import com.example.farmus_application.model.FarmDetail
-
-class FarmDetailRemoteDataSource(private val apiClient: ApiClient): FarmDetailDataSource {
-
-    // apiClient의 함수 결과를 return 하는 FarmDetailDataSource의 suspend 함수를 override
-
-//    override suspend fun getFarmList(): List<FarmDetail> {
-//        return apiClient.getFarmList()
-//    }
-}
