+++ app/src/main/res/layout/rv_local_farm.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="farmList"
-            type="com.example.farmus_application.model.FarmList" />
+            type="com.example.farmus_application.model.farm.FarmList" />
     </data>
 
     <com.google.android.material.card.MaterialCardView
